requirements.txt
langchain>=0.2.0,<0.4.0
langchain-openai>=0.1.0,<0.4.0
langchain-anthropic>=0.1.0,<0.4.0
langchain-community>=0.2.0,<0.4.0
langchain-core>=0.2.0,<0.4.0
openai>=1.0.0
anthropic>=0.25.0
pymupdf>=1.24.0
faiss-cpu>=1.8.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
playwright>=1.44.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.0.0
httpx>=0.27.0
requests>=2.31.0
tqdm>=4.66.0
